[Chore] move `DomEmitter` class to `event.ts`
